assigned_expr: parse iio_trigger_get_irq() correctly
This only affected when iio_trigger_get_irq(), probably because of how
nested inline are handled because then we don't inline
class_mutex_constructor()?

Anyway, Smatch thinks class_mutex_constructor() has two return paths.
Maybe one of the locks is a static variable or something?  It doesn't
matter why, just that it does.  The two returns look like this:

drivers/iio/industrialio-trigger.c | class_mutex_constructor | 68 | 4096-ptr_max[==$0]|        INTERNAL | -1 |                  201 | struct mutex*(*)(struct mutex*) |
drivers/iio/industrialio-trigger.c | class_mutex_constructor | 69 | 7210594643729211392[==$0]|        INTERNAL | -1 |                  201 | struct mutex*(*)(struct mutex*) |

The actual function looks like this:

struct mutex *class_mutex_constructor(struct mutex * _T)
{
    struct mutex *t = ({ mutex_lock_nested(_T, 0); _T; });
    return t;
}

So it returns whatever we pass to it, which the DB records as "==$0".

For this situation we want to change:

	p = class_mutex_constructor(&lock);

Into:

	p = &lock;

So we create a fake "p = &lock;" assignment, parse it, then parse the
p = class_mutex_constructor() assignment and ignore that.  The problem is
that since this was getting split into two returns we parsed the first
return and ignored the second return as well as the call.  We should
instead use get_unfaked_call() and only ignore the call.

The next problem is that when you merge the two returns together it says
&merged.  Since the return is the same in both places, we can squish them
together.

I also made ignored_expr and __in_fake_parameter_assign as function data
so that inline would be parsed better theoretically.

+++ smatch.h
@@ -686,7 +686,7 @@ int in_condition(void);
 /* smatch_flow.c */
 
 extern int __in_fake_assign;
-extern int __in_fake_parameter_assign;
+extern unsigned long __in_fake_parameter_assign;
 extern int __in_fake_struct_assign;
 extern int __in_buf_clear;
 extern int __in_fake_var_assign;
